[WIP][FEATURE] Content Type: FileType
The table sys_file_reference of TYPO3 is special. It has a type field,
which references a value of another table "sys_file". As these values
are fixed, it is not possible to extend it with custom types. This
means the only possibility is to override the types "showitem" section.

Some adjustments on various places are needed to make this work:
- FileField is treated as if it has no typeField
- An exception is made to override the types section
- Append the hidden palette filePalette

todo: Think about a clever syntax for local overrides with
overrideChildTca.

Fixes: #76